reabrindo vscode
